requirements-pc.txt
Keras==2.1.6
numpy==1.14.3
matplotlib==2.2.2
scipy==1.1.0
opencv_contrib_python==3.4.3.18
Flask_Testing==0.7.1
requests==2.19.1
Flask_Cors==3.0.6
Flask==1.0.2
SpeechRecognition==3.8.1
Pillow==5.3.0
protobuf==3.6.1
mock==2.0.0
scikit_learn==0.20.1
google-cloud-translate==1.3.1
mock==2.0.0
Unidecode==1.0.23
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Flask-Login==0.4.1
Flask-Migrate==2.3.1
tensorflow-gpu==1.8.0
azure-cognitiveservices-search-imagesearch==1.0.0